requirements-qa.txt
a2wsgi==1.10.8
anyio
browser-use
hypercorn==0.17.3
ddgs>=9.5.0
lru-dict==1.3.0
fake-useragent==1.5.1
annoy>=1.17
aiohttp>=3.10
googlesearch-python==1.2.3
beautifulsoup4==4.12.3
nest-asyncio==1.6.0
litellm==1.74
langchain-core
langchain
sentence-transformers==3.0.1
cryptography==43.0.0
openai-whisper==20240930
GitPython==3.1.43
faiss-cpu==1.11.0
sqlite-utils>=3.36
fastmcp==2.3.4
fasta2a==0.5.0
PyYAML>=6.0
prometheus_client>=0.22
fastapi>=0.115
flask[async]==3.0.3
flask-basicauth==0.2.0
flaredantic==0.1.4
inputimeout==1.0.4
kokoro>=0.9.2
simpleeval==1.0.3
langchain-community
langchain-unstructured[all-docs]
lxml_html_clean==0.3.1
markdown==3.7
mcp==1.9.0
newspaper3k==0.2.8
pexpect==4.9.0
playwright==1.52.0
ptyprocess==0.7.0
pypdf==4.3.1
python-dotenv==1.1.0
pytz==2024.2
tiktoken==0.8.0
unstructured[all-docs]==0.16.23
unstructured-client==0.31.0
webcolors==24.6.0
crontab==1.0.1
markdownify==1.1.0
pymupdf==1.25.3
pytesseract==0.3.13
pdf2image==1.17.0
pathspec>=0.12.1
psutil>=7.0.0
soundfile==0.13.1
pytest-asyncio>=0.23
hypothesis>=6.100
pytest-cov>=5.0
